+++ tokenizer-api/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "tantivy-tokenizer-api"
+version = "0.1.0"
+edition = "2021"
+description = "Tokenizer API of tantivy"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0.152", features = ["derive"] }
